requirements.txt
streamlit==1.37.1
openai==1.56.1
python-docx
PyPDF2